hopefully reduce requestqueue diff
